#19 feat: Phase 6 완료 - 설정 파일 분리 및 main.py 완전 정리
✅ 설정 시스템 구현:
- app/config/settings.py: 중앙화된 설정 관리
- 환경변수 기반 설정 (KAKAO_API_KEY, BOT_ID 등)
- 로깅 설정, 스케줄링 설정, 알림 설정 등 포함

🔧 main.py 대규모 리팩토링:
- 2,407줄 → 112줄로 대폭 간소화
- main_old.py로 기존 파일 백업
- Router-Service-Repository 패턴 완전 적용
- 깔끔한 FastAPI 앱 구조 (title, version, description)

🏗️ 아키텍처 완성:
- Models (app/models/): Pydantic 데이터 모델
- Database (app/database/): 연결 및 스키마 관리
- Utils (app/utils/): 지리계산, 스케줄러 유틸
- Services (app/services/): 비즈니스 로직 레이어
- Routers (app/routers/): API 엔드포인트 그룹화
- Config (app/config/): 설정 관리

✅ 모듈화 프로젝트 완료!
